refactor: move wagmi config to a separate file
